Project setup is done
